fix: remove .__rc file breaking Nix's use of libarchive on OSX
Nix uses libarchive to bundle the source code in preparation for building
the gitolite package, and on OSX that interprets the `.__rc` file as defining
extended attributes for a `_rc` file, if in fact that existed. The result
is that the resulting narinfo checksums do not match on OSX because this file
is effectively missing when the calculation is performed.

The git commit history reveals this file was created with a patch involving
mirroring, but I don’t see any mention of .__rc or the command alias `glp`
that it seems to be defining. Looking for mentions of /tmp/gitolite.git
would imply that’s just a place where sitaramc likes to keep a bare repo
copy of the source code for convenience, so I imagine this file was just
part of his developer environment setup and not required for the operation
of gitolite itself.

Removing the file (in our fork) to address this issue of building on OSX.

More background on these Apple extended attribute files:
  https://en.wikipedia.org/wiki/AppleSingle_and_AppleDouble_formats