feat: Get user roles from GL_ROLES env var if set
Previously, we granted a user CREATOR permission based on their username
matching the floxmeta repo they were trying to access. We now
preferentially check the GL_ROLES environment variable, which is
expected to be set by gitolite-server. This allows the server, which has
access to the token and future authorization service, to more explicitly
control the permissions at the gitolite level.